Replace custom logger with `log/slog` form stdlib.
The custom logger (`logger.go`) has been removed and replaced with the `slog` logging library. Changes were made to the function `WithLogger` in `option.go` and the `Cron` structure in `cron.go` to incorporate this new logger. Also, the tests for verbose logging in `option_test.go` were removed since it is no longer relevant with the new logging system.